--- src/draw.ts
@@ -1,1 +1,0 @@
-// This file implements the drawing functions for the language.
